feat(textfield): add event handlers when press enter and esc keys
